Rename heap-profiler module to heap-dumper
